Task class: add assert statements in deserialize
Deserialize method just assumes that the correct
number of arguments are provided to correctly
deserialize and create a new Task

This could result in Array Access issues if
the number of tokens is not equal to the
number of arguments needed to create the
new task.

Adding assert statements at least emphasizes
how many tokens should the deserialized String
should have in order to create the Task.

Let's add assert statements to the deserialization
function in the Task class before Task creation.